Fatemeh Davoudi implementation of the symbolic model reduction algorithm For detailed description see http://www.ep.liu.se/ecp/043/090/ecp09430136.pdf The algorithm is implemented in the cpp runtime in Core/ReduceDAE At the moment for the ranking method the residuen ranking is implement in class Core/ReduceDAE/Ranking.cpp The reduction mehtods cancel terms, linearize terms and substitute terms are implemented in class Core/ReduceDAE/Reduction.cpp The label algorithms for cancel terms, linearize terms and substitute terms are implemented in OMCompiler/Compiler/SimCode/ReduceDAE.mo The method can be used with: setCommandLineOptions("+simCodeTarget=Cpp --labeledReduction --disableExtraLabeling -d=writeToBuffer"); loadFile("your model file"); simulate(yourmodel);
added omc c-api communication project to the cpp folder. The c- api is build with the cpp runtime for mingw and gcc. The c-api is used by the reduce dae algorithm to start labeling and reduce terms

adapted reduce dae build for cpp runtime build
added omc C-API library to cpp runtime
added build for C-API library to cpp runtime build
fixed includes
removed unused files
reorganized some code parts

added test for c-api

moved in LoadCompilerSources the include order for ReduceDAE.mo to the backend part

fix in Flags.mo, two Flags used the same id

only build reduce dae for mingw

merged in cpp template model reduction code generation

added missing interface methods

fix for compiler errors whem getDimRHS return nothing

change formtat ReduceDAE.mo to utf8

moved function getNumContinuousEquations to SimCodeUtil

use std chrono instead of boost time library

added sundials dependicies to runtimeCPP make target

fix for wrong type for timeout in SimSettings

fix for wrong reduceDAE includes, if reduceDAE flag is disabled

fix for some tests, where  der variables are filtered in outputfike

fix for missing IDEAReduce include in generated code

Belonging to [master]:
  - OpenModelica/OMCompiler#1973

+++ SimulationRuntime/cpp/Include/Core/DataExchange/HistoryImpl.h
@@ -76,7 +76,7 @@ public:
     return ResultsPolicy::size();
   }
 
-  unsigned long getDimRe()
+ int getDimRe()
   {
     throw ModelicaSimulationError(DATASTORAGE,"getDimRe not implemented yet");
   }
